move providers to components/providers
