+++ plopfile.js
@@ -1,11 +1,10 @@
 import fs from 'fs'
 
-// @plop-main-page-import - Please don't delete. Comments for automatic addition by plop.
 const APPEND_PATTERN = /\/\/ @plop - Please don't delete. Comments for automatic addition by plop./
-const APPEND_PATTERN_MAIN_ROUTE =
-	/\/\/ @plop-main-route - Please don't delete. Comments for automatic addition by plop./
-const APPEND_PATTERN_MAIN_PAGE_IMPORT =
-	/\/\/ @plop-main-page-import - Please don't delete. Comments for automatic addition by plop./
+const APPEND_PATTERN_APP_ROUTE =
+	/\/\/ @plop-route - Please don't delete. Comments for automatic addition by plop./
+const APPEND_PATTERN_APP_ROUTE_IMPORT =
+	/\/\/ @plop-route-import - Please don't delete. Comments for automatic addition by plop./
 
 const componentKinds = fs.readdirSync('src/components').map((kind) => {
 	return {
@@ -53,15 +52,15 @@ export default function (plop) {
 			},
 			{
 				type: 'append',
-				path: 'src/main.tsx',
-				templateFile: 'plop-templates/main/route.hbs',
-				pattern: APPEND_PATTERN_MAIN_ROUTE,
+				path: 'src/components/providers/CustomRouterProvider/index.tsx',
+				templateFile: 'plop-templates/route/route.hbs',
+				pattern: APPEND_PATTERN_APP_ROUTE,
 			},
 			{
 				type: 'append',
-				path: 'src/main.tsx',
-				templateFile: 'plop-templates/main/import-page.hbs',
-				pattern: APPEND_PATTERN_MAIN_PAGE_IMPORT,
+				path: 'src/components/providers/CustomRouterProvider/index.tsx',
+				templateFile: 'plop-templates/route/import.hbs',
+				pattern: APPEND_PATTERN_APP_ROUTE_IMPORT,
 			},
 		],
 	})
