+++ src/components/providers/CustomThemeProvider/index.tsx
@@ -1,0 +1,10 @@
+import { theme } from '@/theme'
+import { ChakraProvider } from '@chakra-ui/react'
+
+type CustomThemeProviderProps = {
+	children: React.ReactNode
+}
+
+export function CustomThemeProvider({ children }: CustomThemeProviderProps) {
+	return <ChakraProvider theme={theme}>{children}</ChakraProvider>
+}
