+++ src/theme/colors/index.ts
@@ -1,2 +1,2 @@
 // https://chakra-ui.com/docs/styled-system/customize-theme#customizing-theme-tokens
-export const CustomColorOptions = {}
+export const customColorOptions = {}
