+++ src/theme/styles/index.ts
@@ -1,5 +1,5 @@
 // https://chakra-ui.com/docs/styled-system/customize-theme#customizing-global-styles
-export const CustomStyleOptions = {
+export const customStyleOptions = {
 	global: {
 		body: {},
 	},
